Test and main code for delete with user name field.
https://github.com/marcocalamai/to-do-list/issues/8

+++ todo-list/src/main/java/org/marco/calamai/todolist/services/ToDoService.java
@@ -8,6 +8,7 @@ import java.util.Optional;
 
 import org.marco.calamai.todolist.exceptions.InvalidTimeException;
 import org.marco.calamai.todolist.exceptions.ToDoNotFoundException;
+import org.marco.calamai.todolist.exceptions.WrongUsernameException;
 import org.marco.calamai.todolist.model.ToDo;
 import org.marco.calamai.todolist.repositories.mongo.ToDoMongoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +19,7 @@ public class ToDoService {
 	
 	public static final String DATE_IS_BEFORE_TODAY = "Date not valid. It has passed!";
 	public static final String TO_DO_NOT_FOUND = "ToDo not found!";
+	public static final String WRONG_USERNAME = "Wrong username!";
 
 	
 	@Autowired
@@ -62,9 +64,12 @@ public class ToDoService {
 		throw new ToDoNotFoundException(TO_DO_NOT_FOUND);
 	}
 
-	public ToDo deleteToDoById(BigInteger id) {
+	public ToDo deleteToDoById(BigInteger id, String username) {
 		Optional<ToDo> toDo = toDoMongoRepository.findById(id);
 		if (toDo.isPresent()){
+			if (!toDo.get().getUser().equals(username) ) {
+				throw new WrongUsernameException(WRONG_USERNAME);
+			}
 			toDoMongoRepository.deleteById(id);
 			return toDo.get();
 		}
